refactor: extract _validate_motif to api/routes/_common.py
