+++ api/routes/tasks.py
@@ -1,12 +1,6 @@
 from typing import Literal
 from fastapi import APIRouter, Depends, HTTPException, Query, Request
 from pydantic import BaseModel
-
-
-class MoveTaskBody(BaseModel):
-    date: str
-    anchor_id: str
-    position: int | None = None
 import asyncpg
 from db.pg_queries import (
     patch_task_fields, move_task_atomic,
@@ -18,18 +12,16 @@ from db.pg_queries import (
 )
 from db.pg_queries.tasks import set_task_rrule, delete_anchor_occurrence, truncate_anchor_series
 from db.pool_middleware import get_db_conn
-from db.pg_queries._motif import VALID_MOTIFS
+from api.routes._common import _validate_motif
 from api.auth import auth_dependency
 
 router = APIRouter()
 
 
-def _validate_motif(body: dict) -> None:
-    if "motif" in body and body["motif"] not in VALID_MOTIFS:
-        raise HTTPException(
-            status_code=422,
-            detail=f"Invalid motif: {body['motif']!r}. Must be one of {sorted(VALID_MOTIFS)}",
-        )
+class MoveTaskBody(BaseModel):
+    date: str
+    anchor_id: str
+    position: int | None = None
 
 
 class TaskRruleBody(BaseModel):
